Implement RedNotebook HTML importing (#1026)
* RedNotebook importing working

Also split HTML parsing into ct_parser_html.cc

* Added test data to /tests/data

* Add ct_parser_html.cc to git

* Fix license naming

* Remove ct_parser_html

* Add moved HTML code from ct_parser

* Add code for CtRedNotebookParser
