+++ future/src/ct/CMakeLists.txt
@@ -57,6 +57,7 @@ set(CT_SHARED_FILES
     ct_widgets.cc
     ct_parser_text.cc
     ct_parser_md.cc
+    ct_parser_html.cc
     ct_parser.cc
     ct_process.cc
     ct_filesystem.cc
