+++ future/src/ct/ct_actions.h
@@ -424,6 +424,7 @@ public:
     void import_nodes_from_keepnote_directory() noexcept;
     void import_nodes_from_mempad_file() noexcept;
     void import_nodes_from_leo_file() noexcept;
+    void import_nodes_from_rednotebook_html() noexcept;
 
 private:
     // helper for export actions
