+++ future/src/ct/ct_actions_import.cc
@@ -156,6 +156,16 @@ void CtActions::import_nodes_from_leo_file() noexcept {
     }
 }
 
+void CtActions::import_nodes_from_rednotebook_html() noexcept {
+    try {
+        CtRedNotebookImporter importer{_pCtMainWin->get_ct_config()};
+        _import_from_file(&importer);
+    } catch(const std::exception& e) {
+        spdlog::error("Exception caught while importing from Leo: {}", e.what());
+    }
+}
+
+
 void CtActions::_import_from_file(CtImporterInterface* importer)
 {
     CtDialogs::file_select_args args(_pCtMainWin);
