+++ future/tests/data/rednotebook_data.html
@@ -1,0 +1,95 @@
+<?xml version="1.0"
+      encoding="UTF-8"
+?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml">
+<head><meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+<meta name="generator" content="http://txt2tags.org" />
+<style type="text/css">
+    :root {
+        color-scheme: light dark;
+        --fgcolor: rgb(49,54,61);
+        --bgcolor: rgb(248,248,248);
+        --dark-link-color: rgb(0, 188, 212);
+        --light-link-color: rgb(0, 0, 238);
+    }
+    a { color: var(--light-link-color); }
+    body {
+        font-family: Ubuntu, mingliu, MS Mincho, sans-serif;
+        background: var(--bgcolor);
+        color: var(--fgcolor);
+    }
+    @media (prefers-color-scheme: dark) {
+        a { color: var(--dark-link-color); }
+    }
+    <!-- Don't split last line between pages.
+         This fix is only supported by Opera -->
+    p {
+        page-break-inside: avoid;
+    }
+    blockquote {
+        margin: 1em 2em;
+        border-left: 2px solid #999;
+        font-style: oblique;
+        padding-left: 1em;
+    }
+    blockquote:first-letter {
+        margin: .2em .1em .1em 0;
+        font-size: 160%;
+        font-weight: bold;
+    }
+    blockquote:first-line {
+        font-weight: bold;
+    }
+    table {
+        border-collapse: collapse;
+    }
+    td, th {
+        <!--border: 1px solid #888;--> <!--Allow tables without borders-->
+        padding: 3px 7px 2px 7px;
+    }
+    th {
+        text-align: left;
+        padding-top: 5px;
+        padding-bottom: 4px;
+        background-color: #aaa;
+        color: #ffffff;
+    }
+    hr.heavy {
+        height: 2px;
+        background-color: black;
+    }
+</style>
+</head>
+<body>
+
+<div class="header" id="header">
+</div>
+
+<div class="body" id="body">
+<p>
+<span id="2020-07-15"></span>
+</p>
+
+<h1>Wednesday, 15/07/20</h1>
+
+<p>
+Some Text<i>Italic</i>
+</p>
+<p>
+<b>Bold</b>
+</p>
+<p>
+<span id="2020-07-16"></span>
+</p>
+
+<h1>Thursday, 16/07/20</h1>
+
+<p>
+Second day
+</p>
+</div>
+
+<!-- xhtml code generated by txt2tags 2.6 (http://txt2tags.org) -->
+<!-- cmdline: txt2tags  -->
+</body></html>
